// src/types/TwitchData.tsx
export interface Schedule {
  segments: Segment[];
  broadcaster_id: string;
  broadcaster_name: string;
  broadcaster_login: string;
  vacation: Vacation;
}

export interface Vacation {
  start_time: string;
  end_time: string;
}

export interface Segment {
  id: string;
  start_time: string;
  end_time: string;
  title: string;
  canceled_until: string;
  category: {
    id: string;
    name: string;
  };
  is_recurring: boolean;
}